refactor(core): move to seperate files treecrdt helper functions and types
